Light (#1)
* The PSI update

Added an overlayed \psi to light.svg

* The CODE update

Here we take the idea of the underlayed log file (refer to commit 6d102f3) and tweak it a little: a portion of the DMFT-ED code, instead of a log.

Suggested by @aamaricci.

* No more PSI

Deleted the overlayed \PSI: the curves are not symmetric enough for that. It might also be an overall bad idea. Let's just take it out for now.