Beyond dark/light
Here we introduce some branded color-schemes. The idea is that the logo might become color agnostic: it changes scheme dependent on context (e.g. a julia-lang API repository will feature the julia-colors logo).

NB. we should understand if some of these color schemes are proprietary; jl colors are for sure free to use, since almost every big julia package uses it for its logo; google, microsoft and slack are in doubt for now, but I have a good feeling about this. If I'd happen to be wrong we'll move away from every copyrighted scheme, of course.