Brand colors for ED
Ref. to commit 6f40b61

An additional note: always modify the alpha on the layers, not on the colors, so to preserve the hex values as written in the rgba.hex file (i.e. everything should always end in `ff`). To comply with this, dark and light files have also been modified.