Freezing the branch
This branch implements an idea proposed and promoted by @frapao, namely to move away from the overall "Ψ-shape" embracing instead an "ED", acronym for Exact Diagonalization, intended as a more representative keyword for the whole project. Given the current state of the codebases and the roadmap for future additions the idea has been dropped, as the scope appears to be too tight.

Nevertheless we leave it here as it looks quite good and could still be useful in some form for the future.

So, please:

- Don't delete the branch.
- Don't merge to master.